Build node/browser bundles
