+++ webpack.config.js
@@ -1,42 +1,34 @@
-const config = {
-  mode: "production",
-  entry: {
-    ['brewcalc.min']: './src/index.js',
-    import: './src/importFromBeerXml.js'
+const config = [
+  {
+    mode: "production",
+    output: {
+      library: "brewcalc",
+      libraryTarget: "umd",
+      umdNamedDefine: true,
+      path: __dirname + "/lib",
+      filename: "brewcalc.min.js"
+    },
+    optimization: {
+      minimize: true
+    },
+    module: {
+      rules: require("./webpack.config.rules")
+    }
   },
-  output: {
-    library: "brewcalc",
-    libraryTarget: "umd",
-    umdNamedDefine: true,
-    path: __dirname + "/lib",
-    filename: "[name].js"
-  },
-  optimization: {
-    minimize: true
-  },
-  module: {
-    rules: [
-      {
-        test: /\.js?$/,
-        enforce: "pre",
-        loader: "prettier-loader",
-        options: {
-          parser: "babel",
-          singleQuote: true,
-          semi: false,
-          jsxBracketSameLine: true,
-          tabWidth: 2
-        }
-      },
-      {
-        test: /\.js$/,
-        exclude: /node_modules/,
-        use: {
-          loader: "babel-loader"
-        }
-      }
-    ]
+  {
+    mode: "development",
+    output: {
+      library: "brewcalc",
+      libraryTarget: "umd",
+      umdNamedDefine: true,
+      path: __dirname + "/lib",
+      filename: "brewcalc.js"
+    },
+    devtool: "source-map",
+    module: {
+      rules: require("./webpack.config.rules")
+    }
   }
-}
+]
 
 module.exports = config
